Beta 1.2 update changes...

+++ src/inventory.cpp
@@ -224,7 +224,7 @@ bool Inventory::readRecipe(std::string recipeFile)
   return true;
 }
 
-bool Inventory::windowClick(User *user,int8_t windowID, int16_t slot, int8_t rightClick, int16_t actionNumber, int16_t itemID, int8_t itemCount,int8_t itemUses)
+bool Inventory::windowClick(User *user,int8_t windowID, int16_t slot, int8_t rightClick, int16_t actionNumber, int16_t itemID, int8_t itemCount,int16_t itemUses)
 {  
   //Ack
   user->buffer << (int8_t)PACKET_TRANSACTION << (int8_t)windowID << (int16_t)actionNumber << (int8_t)1;
@@ -584,7 +584,7 @@ bool Inventory::windowClick(User *user,int8_t windowID, int16_t slot, int8_t rig
           (*otherUsers)[i]->buffer << (int8_t)PACKET_SET_SLOT << (int8_t)windowID << (int16_t)slot << (int16_t)slotItem->type;
           if(slotItem->type != -1)
           {
-            (*otherUsers)[i]->buffer << (int8_t)slotItem->count << (int8_t)slotItem->health;
+            (*otherUsers)[i]->buffer << (int8_t)slotItem->count << (int16_t)slotItem->health;
           }
         }
       }
@@ -601,7 +601,7 @@ bool Inventory::windowClick(User *user,int8_t windowID, int16_t slot, int8_t rig
             (*otherUsers)[i]->buffer << (int8_t)PACKET_SET_SLOT << (int8_t)windowID << (int16_t)slot << (int16_t)slotItem->type;
             if(slotItem->type != -1)
             {
-              (*otherUsers)[i]->buffer << (int8_t)slotItem->count << (int8_t)slotItem->health;
+              (*otherUsers)[i]->buffer << (int8_t)slotItem->count << (int16_t)slotItem->health;
             }
           }
         }
@@ -619,7 +619,7 @@ bool Inventory::windowClick(User *user,int8_t windowID, int16_t slot, int8_t rig
             (*otherUsers)[i]->buffer << (int8_t)PACKET_SET_SLOT << (int8_t)windowID << (int16_t)slot << (int16_t)slotItem->type;
             if(slotItem->type != -1)
             {
-              (*otherUsers)[i]->buffer << (int8_t)slotItem->count << (int8_t)slotItem->health;
+              (*otherUsers)[i]->buffer << (int8_t)slotItem->count << (int16_t)slotItem->health;
             }
           }
         }
@@ -658,7 +658,7 @@ bool Inventory::windowOpen(User *user, int8_t type, int32_t x, int32_t y, int32_
             if(chunk->chests[i]->items[j].type != -1)
             {
               user->buffer << (int8_t)PACKET_SET_SLOT << (int8_t)WINDOW_CHEST << (int16_t)j << (int16_t)chunk->chests[i]->items[j].type 
-                           << (int8_t)(chunk->chests[i]->items[j].count) << (int8_t)chunk->chests[i]->items[j].health;
+                           << (int8_t)(chunk->chests[i]->items[j].count) << (int16_t)chunk->chests[i]->items[j].health;
             }
           }
           break;
@@ -679,7 +679,7 @@ bool Inventory::windowOpen(User *user, int8_t type, int32_t x, int32_t y, int32_
             if(openWorkbenches[i]->workbench[j].type != -1)
             {
               user->buffer << (int8_t)PACKET_SET_SLOT << (int8_t)WINDOW_WORKBENCH << (int16_t)j << (int16_t)openWorkbenches[i]->workbench[j].type 
-                           << (int8_t)(openWorkbenches[i]->workbench[j].count) << (int8_t)openWorkbenches[i]->workbench[j].health;
+                           << (int8_t)(openWorkbenches[i]->workbench[j].count) << (int16_t)openWorkbenches[i]->workbench[j].health;
             }
           }
           break;
@@ -699,7 +699,7 @@ bool Inventory::windowOpen(User *user, int8_t type, int32_t x, int32_t y, int32_
             if(chunk->furnaces[i]->items[j].type != -1)
             {
               user->buffer << (int8_t)PACKET_SET_SLOT << (int8_t)WINDOW_FURNACE << (int16_t)j << (int16_t)chunk->furnaces[i]->items[j].type 
-                           << (int8_t)(chunk->furnaces[i]->items[j].count) << (int8_t)chunk->furnaces[i]->items[j].health;
+                           << (int8_t)(chunk->furnaces[i]->items[j].count) << (int16_t)chunk->furnaces[i]->items[j].health;
             }
           }
           break;
@@ -757,7 +757,7 @@ bool Inventory::addItems(User *user,int16_t itemID, char count, int16_t health)
     //If slot empty, put item there
     if(slot->type == -1)
     {
-      user->buffer << (int8_t)PACKET_SET_SLOT << (int8_t)0 << (int16_t)(i+9) << (int16_t)itemID << (int8_t)count << (int8_t)health;
+      user->buffer << (int8_t)PACKET_SET_SLOT << (int8_t)0 << (int16_t)(i+9) << (int16_t)itemID << (int8_t)count << (int16_t)health;
       slot->type   = itemID;
       slot->count  = count;
       slot->health = health;
@@ -770,7 +770,7 @@ bool Inventory::addItems(User *user,int16_t itemID, char count, int16_t health)
       //Put to the stack
       if(64-slot->count >= count)
       {
-        user->buffer << (int8_t)PACKET_SET_SLOT << (int8_t)0 << (int16_t)(i+9) << (int16_t)itemID << (int8_t)(slot->count+count) << (int8_t)health;
+        user->buffer << (int8_t)PACKET_SET_SLOT << (int8_t)0 << (int16_t)(i+9) << (int16_t)itemID << (int8_t)(slot->count+count) << (int16_t)health;
         slot->type   = itemID;
         slot->count += count;
         break;
@@ -778,7 +778,7 @@ bool Inventory::addItems(User *user,int16_t itemID, char count, int16_t health)
       //Put some of the items to this stack and continue searching for space
       else if(64-slot->count > 0)
       {
-        user->buffer << (int8_t)PACKET_SET_SLOT << (int8_t)0 << (int16_t)(i+9) << (int16_t)itemID << (int8_t)64 << (int8_t)health;
+        user->buffer << (int8_t)PACKET_SET_SLOT << (int8_t)0 << (int16_t)(i+9) << (int16_t)itemID << (int8_t)64 << (int16_t)health;
         slot->type = itemID;
         slot->count = 64;
         count -= 64-slot->count;
@@ -993,7 +993,7 @@ bool Inventory::setSlot(User *user, int8_t windowID, int16_t slot, int16_t itemI
   user->buffer << (int8_t)PACKET_SET_SLOT << (int8_t)windowID << (int16_t)slot   << (int16_t)itemID;
   if(itemID != -1)
   {
-    user->buffer << (int8_t)count << (int8_t)health;
+    user->buffer << (int8_t)count << (int16_t)health;
   }
 
   return true;
